Restructure the staking page
